pyproject.toml
[project]
name = "coindcx-mcp"
version = "0.1.0"
description = "MCP server for CoinDCX API integration"
authors = [
    {name = "Biswajit Kumar", email = "bksinha4497@gmail.com.com"}
]
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["coindcx_mcp"]

[project.scripts]
coindcx-mcp = "coindcx_mcp.server:main"
